feat : ValidateLotto created
