+++ src/test/java/controller/mock/MockLottoFactory.java
@@ -7,7 +7,6 @@ import model.LottoFactory;
 import java.util.List;
 
 public class MockLottoFactory extends LottoFactory {
-    public boolean checkPriceHigherThanSingleLottoPriceCalled= false;
     public int generateLottoCalledCount = 0;
     public int getLottoCallCount = 0;
 
@@ -25,12 +24,6 @@ public class MockLottoFactory extends LottoFactory {
     public List<Lotto> generateLottoByPrice(int userCashInput) {
         getLottoCallCount++;
         return super.generateLottoByPrice(userCashInput);
-    }
-
-   @Override
-    protected void checkPriceHigherThanSingleLottoPrice(int price) {
-        checkPriceHigherThanSingleLottoPriceCalled = true;
-        super.checkPriceHigherThanSingleLottoPrice(price);
     }
 }
